Add namespaces for event store and projection store adapter modules.
