+++ lib/incident/event_store/in_memory/event.ex
@@ -1,6 +1,6 @@
-defmodule Incident.EventStore.PostgresEvent do
+defmodule Incident.EventStore.InMemory.Event do
   @moduledoc """
-  Defines the data structure for any event for the Postgres adapter.
+  Defines the data structure for any event for the in memory adapter.
 
   All fields are required.
   """
@@ -12,23 +12,19 @@ defmodule Incident.EventStore.PostgresEvent do
           event_type: String.t() | nil,
           version: pos_integer | nil,
           event_date: DateTime.t() | nil,
-          event_data: map | nil,
-          inserted_at: DateTime.t() | nil
+          event_data: map | nil
         }
 
   use Ecto.Schema
   import Ecto.Changeset
 
-  @timestamps_opts [type: :utc_datetime_usec]
-  schema "events" do
-    field(:event_id, :binary_id)
+  embedded_schema do
+    field(:event_id, :string)
     field(:aggregate_id, :string)
     field(:event_type, :string)
     field(:version, :integer)
-    field(:event_date, :utc_datetime_usec)
+    field(:event_date, :utc_datetime)
     field(:event_data, :map)
-
-    timestamps(updated_at: false)
   end
 
   @required_fields ~w(event_id aggregate_id event_type version event_date event_data)a
@@ -39,6 +35,5 @@ defmodule Incident.EventStore.PostgresEvent do
     record
     |> cast(params, @required_fields)
     |> validate_required(@required_fields)
-    |> validate_number(:version, greater_than: 0)
   end
 end
