+++ test/incident/event_store/postgres_adapter_test.exs
@@ -3,10 +3,10 @@ defmodule Incident.EventStore.PostgresAdapterTest do
 
   alias Ecto.UUID
 
-  alias Incident.EventStore.{PostgresAdapter, PostgresEvent, TestRepo}
+  alias Incident.EventStore.{Postgres.Adapter, Postgres.Event, TestRepo}
 
   setup do
-    PostgresAdapter.start_link(repo: TestRepo)
+    Adapter.start_link(repo: TestRepo)
 
     on_exit(fn ->
       Application.stop(:incident)
@@ -29,14 +29,14 @@ defmodule Incident.EventStore.PostgresAdapterTest do
       event_added = %CounterAdded{aggregate_id: @aggregate_id, amount: 5, version: 1}
       event_removed = %CounterRemoved{aggregate_id: @aggregate_id, amount: 4, version: 2}
 
-      assert {:ok, %PostgresEvent{}} = PostgresAdapter.append(event_added)
-      assert {:ok, %PostgresEvent{}} = PostgresAdapter.append(event_removed)
+      assert {:ok, %Event{}} = Adapter.append(event_added)
+      assert {:ok, %Event{}} = Adapter.append(event_removed)
     end
 
     test "does not append a new event into the event store when event is invalid" do
       event = %CounterAdded{aggregate_id: @aggregate_id, amount: 5, version: nil}
 
-      assert {:error, %Ecto.Changeset{valid?: false}} = PostgresAdapter.append(event)
+      assert {:error, %Ecto.Changeset{valid?: false}} = Adapter.append(event)
     end
   end
 
@@ -44,15 +44,14 @@ defmodule Incident.EventStore.PostgresAdapterTest do
     test "returns a list of events for an aggregate id in order" do
       event_added = %CounterAdded{aggregate_id: @aggregate_id, amount: 3, version: 1}
       event_removed = %CounterRemoved{aggregate_id: @aggregate_id, amount: 1, version: 2}
-      PostgresAdapter.append(event_added)
-      PostgresAdapter.append(event_removed)
+      Adapter.append(event_added)
+      Adapter.append(event_removed)
 
-      assert [%PostgresEvent{version: 1}, %PostgresEvent{version: 2}] =
-               PostgresAdapter.get(@aggregate_id)
+      assert [%Event{version: 1}, %Event{version: 2}] = Adapter.get(@aggregate_id)
     end
 
     test "returns an empty list when no events are found" do
-      assert [] = PostgresAdapter.get(UUID.generate())
+      assert [] = Adapter.get(UUID.generate())
     end
   end
 end
