+++ test/incident/event_store/postgres_event_test.exs
@@ -2,7 +2,7 @@ defmodule Incident.EventStore.PostgresEventTest do
   use Incident.RepoCase, async: true
 
   alias Ecto.UUID
-  alias Incident.EventStore.PostgresEvent
+  alias Incident.EventStore.Postgres.Event
 
   @valid_params %{
     event_id: UUID.generate(),
@@ -15,13 +15,13 @@ defmodule Incident.EventStore.PostgresEventTest do
 
   describe "changeset/2" do
     test "returns a valid changeset when all fields are valid" do
-      changeset = PostgresEvent.changeset(%PostgresEvent{}, @valid_params)
+      changeset = Event.changeset(%Event{}, @valid_params)
 
       assert changeset.valid?
     end
 
     test "returns an error when a required field is not present" do
-      changeset = PostgresEvent.changeset(%PostgresEvent{}, %{})
+      changeset = Event.changeset(%Event{}, %{})
 
       refute changeset.valid?
       assert %{event_id: ["can't be blank"]} = errors_on(changeset)
@@ -34,7 +34,7 @@ defmodule Incident.EventStore.PostgresEventTest do
 
     test "returns an error when a field is set with a wront type" do
       invalid_params = Map.merge(@valid_params, %{event_date: "2010-04-11"})
-      changeset = PostgresEvent.changeset(%PostgresEvent{}, invalid_params)
+      changeset = Event.changeset(%Event{}, invalid_params)
 
       refute changeset.valid?
       assert %{event_date: ["is invalid"]} = errors_on(changeset)
